# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API GRATIS",
    "contact": {
      "name": "Akuivan13",
      "url": "https://riskimivan.my.id",
      "whatsapp": "+6289505520763"
    }
  },
  "paths": {
...
    "/api/tiktok": {
      "get": {
        "tags": ["Downloader"],
        "summary": "Unduh Video TikTok",
        "description": "Endpoint untuk mengunduh video TikTok",
               "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "Url Tiktok Downloader",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/tiktok": {
      "get": {
        "tags": ["Downloader"],
        "summary": "Download video dari TikTok",
        "description": "Endpoint untuk download video dari TikTok",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL video di TikTok"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}